tests: Adds mock patch autospec fixture
When mocking something, there's the possibility to reference a
property that does not exist, or call a method with invalid arguments,
but the unit tests to keep passing, resulting in false positives.

We've seen this issue several times in OpenStack, and we have created
and started using a mock patch fixture (can be seen in OpenStack Nova's test.py)
which enables autospec by default.

This guarantees 2 things: that the thing we're patching exists, and if it's
callable, that the calls respect the callable's signature (e.g.: no unknown
arguments can be used).